Add apache2db script
apache2db.pl extracts the relevant fields from a set of apache logs and inserts
them into a PostgreSQL table with the schema given in apache2db.sql.  To avoid
duplicates, the script assumes that the table has a unique key on logtime.  The
database connection info is taken from apache2db.conf.