Use stored dates to skip apache2db reprocessing
If apache2db.pl was previously run, the uplog table will already contain rows.
Use the maximum date previously stored to skip retrying older logs.  The "ON
CONFLICT" clause will prevent these from going in, but it should be faster to
skip them locally instead of sending to the server and letting it find a
conflict.

A new LOGOVERLAP option is added to apache2db.conf that controls how many
minutes of overlap are allowed.  It is set to 90 minutes by default to account
for DST transition oddities.