Convert missing_dsl_checkins to read from database
Now that apache2db can parse logs and store them in the database, it doesn't
make sense for missing_dsl_checkins to re-parse the same log files.  Switch it
to read from the database.  It uses the same config for the database connection
info and assumes the same schema for the uplog table.

Since the number of log entries to read is no longer controlled externally,
also add a --days option while we're at it.  The default is 31 days to
approximately match the previous amount of log lines that were being fed in.